+++ Sources/AjarCore/Edit/ClipCopyingEditCommands.swift
@@ -15,7 +15,9 @@ extension EditReducer {
         effects: ClipEffects? = nil,
         effectsAnimation: AnimatableClipEffects? = nil,
         audioMix: ClipAudioMix? = nil,
-        speed: RationalValue? = nil
+        speed: RationalValue? = nil,
+        reverse: Bool? = nil,
+        freezeFrame: Bool? = nil
     ) -> Clip {
         let replacementTransform = transform ?? clip.transform
         let replacementAnimation = transformAnimation
@@ -41,7 +43,9 @@ extension EditReducer {
             effects: replacementEffects,
             effectsAnimation: replacementEffectsAnimation,
             audioMix: audioMix ?? clip.audioMix,
-            speed: speed ?? clip.speed
+            speed: speed ?? clip.speed,
+            reverse: reverse ?? clip.reverse,
+            freezeFrame: freezeFrame ?? clip.freezeFrame
         )
     }
 }